requirements.txt
gradio>=4.0.0
transformers>=4.35.0
torch>=2.0.0
pillow>=10.0.0
python-dotenv>=1.0.0
spaces>=0.1.0
accelerate>=0.24.0
spaces